Add function to check for sub items in a category not owned
Also, function changes to support adding user_id to categories
and Items.

The function to support looking up an items that are part of a
category that is not owned by the owner of the category, is really
just to help deal with the problem that I let users create and own
categories and items, and don't prevent accounts adding items to
categories that they don't own.

In the real world, this would probably need a totally different
aproach. For instance.  I might need a lock on creating items
in the category while the checking and deleting were happening.